backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.5

# Core ML stack (Py3.9-friendly)
numpy==1.26.4
torch==2.2.2
faiss-cpu==1.8.0
transformers==4.41.2
sentence-transformers==2.6.1
accelerate==0.28.0

python-dateutil==2.9.0.post0
tiktoken==0.6.0
pypdf==4.3.1
python-docx==1.1.2

# OpenAI client (pin 1.x) + compatible httpx
openai==1.51.0
httpx==0.27.*

python-dotenv==1.0.1

